Move Node Callbacks into a Hook
